稍作修改

+++ system/projects.py
@@ -43,6 +43,12 @@ class Projects(Resource):
     @access_required(role_dict["product"])
     def delete(self, project_id):
         db = DB()
+        try:
+            status, message = update_group_for_delete_project(project_id)
+        except Exception as e:
+            logger.info('Exception:'+str(e))
+        if status is not True:
+            return {"status": False, "message": message}, 500
         status, result = db.delete_by_id("projects", project_id)
         db.close_mysql()
         if status is not True:
@@ -50,13 +56,6 @@ class Projects(Resource):
             return {"status": False, "message": result}, 500
         if result is 0:
             return {"status": False, "message": "%s does not exist" % project_id}, 404
-        try:
-            status, message = update_group_for_delete_project(project_id)
-        except Exception as e:
-            logger.info('Exception:'+str(e))
-
-        if status is not True:
-            return {"status": False, "message": message}, 500
         return {"status": True, "message": ""}, 200
 
     @access_required(role_dict["product"])
